refactor: move updateWhitelist to whitelist.go
